+++ mdentropy/core/__init__.py
@@ -1,0 +1,2 @@
+from .entropy import *
+from .information import *
