+++ mdentropy/metrics/__init__.py
@@ -1,0 +1,2 @@
+from .mutinf import *
+from .tent import *
