+++ mdentropy/tests/test_tent.py
@@ -2,40 +2,27 @@ import os
 import shutil
 import tempfile
 
-from mdentropy.mutinf import DihedralMutualInformation
-from mdentropy.tent import DihedralTransferEntropy
-
+import numpy as np
 from numpy.testing import assert_almost_equal as eq
 
-import mdtraj as md
+from mdentropy.metrics import DihedralTransferEntropy
+from mdentropy.core import cmi, ce, ncmi
 
+import mdtraj as md
 from msmbuilder.example_datasets import FsPeptide
 
 
-def test_fs_mi():
+def test_ncmi():
+    A = np.random.uniform(low=-180., high=180, size=1000)
+    B = np.random.uniform(low=-180., high=180, size=1000)
+    C = np.random.uniform(low=-180., high=180, size=1000)
 
-    cwd = os.path.abspath(os.curdir)
-    dirname = tempfile.mkdtemp()
-    FsPeptide(dirname).get()
+    NCMI = ncmi(30, A, B, C)
 
-    try:
-        os.chdir(dirname)
-
-        top = md.load(dirname + '/fs_peptide/fs-peptide.pdb')
-        traj = md.load(dirname + '/fs_peptide/trajectory-1.xtc',
-                       stride=10, top=top)
-
-        mi = DihedralMutualInformation()
-        M = mi.partial_transform(traj)
-
-        eq(M[0, 1], M[1, 0])
-
-    finally:
-        os.chdir(cwd)
-        shutil.rmtree(dirname)
+    eq(NCMI, cmi(30, A, B, C)/ce(30, A, C), 6)
 
 
-def test_fs_tent():
+def test_dihedral_tent():
 
     cwd = os.path.abspath(os.curdir)
     dirname = tempfile.mkdtemp()
